Better arrangement of tests
