+++ datalad_next/constraints/tests/test_special_purpose.py
@@ -1,3 +1,4 @@
+from io import StringIO
 import pathlib
 import pytest
 
@@ -8,6 +9,8 @@ from ..basic import (
     EnsureStr,
     NoConstraint,
 )
+from ..compound import EnsureGeneratorFromFileLike
+from ..formats import EnsureJSON
 from ..git import (
     EnsureGitRefName,
 )
@@ -128,3 +131,33 @@ def test_EnsureParameterConstraint():
             pytest.deprecated_call():
         EnsureParameterConstraint.from_parameter(
             Parameter(), 2, item_constraint='unknown')
+
+
+nested_json = """\
+{"name": "Alexa", "wins": [["two pair", "4♠"], ["two pair", "9♠"]]}
+"""
+nested_json_decoded = {
+    "name": "Alexa",
+    "wins": [["two pair", "4♠"],
+             ["two pair", "9♠"]],
+}
+invalid_json = """\
+{"name": BOOM!}
+"""
+
+
+def test_EnsureJSONLines():
+    constraint = EnsureGeneratorFromFileLike(EnsureJSON())
+
+    assert 'items of type "JSON" read from a file-like' \
+        ==  constraint.short_description()
+
+    # typical is "object", but any valid JSON value type must work
+    assert list(constraint(StringIO("5"))) == [5]
+    # unicode must work
+    uc = "ΔЙקم๗あ"
+    assert list(constraint(StringIO(f'"{uc}"'))) == [uc]
+    assert list(constraint(StringIO(nested_json))) == [nested_json_decoded]
+
+    with pytest.raises(ValueError) as e:
+        list(constraint(StringIO(f'{nested_json}\n{invalid_json}')))
